+++ app/Http/Controllers/Api/Admin/StaffUserController.php
@@ -125,4 +125,6 @@ class StaffUserController extends Controller
 
         return ResponseService::success('Password updated successfully.');
     }
+
+
 }
